// src/api/generated/model/addUserAddressCommand.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * ECommerce API
 * ECommerce API with OpenIddict Authentication and Versioning
 * OpenAPI spec version: v1
 */

export interface AddUserAddressCommand {
  userId?: string;
  /** @nullable */
  label?: string | null;
  /** @nullable */
  street?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  zipCode?: string | null;
  /** @nullable */
  country?: string | null;
  isDefault?: boolean;
}